utils MakeJson2ProtoConfig
