+++ ydb/mvp/meta/mvp.cpp
@@ -8,15 +8,14 @@
 #include <ydb/library/actors/http/http_proxy.h>
 #include <ydb/library/actors/interconnect/poller/poller_actor.h>
 #include <ydb/library/actors/protos/services_common.pb.h>
-#include <ydb/library/yaml_json/yaml_to_json.h>
 
 #include <ydb/mvp/core/core_ydb.h>
 #include <ydb/mvp/core/core_ydbc.h>
 #include <ydb/mvp/core/protos/mvp.pb.h>
+#include <ydb/mvp/core/utils.h>
 
 #include <google/protobuf/text_format.h>
 #include <library/cpp/deprecated/atomic/atomic.h>
-#include <library/cpp/protobuf/json/json2proto.h>
 #include <yaml-cpp/yaml.h>
 
 #include <util/datetime/base.h>
@@ -29,22 +28,6 @@ using namespace NMVP;
 
 NMVP::TMVP* NMVP::InstanceMVP;
 namespace {
-
-NProtobufJson::TJson2ProtoConfig MakeJson2ProtoConfig() {
-    return NProtobufJson::TJson2ProtoConfig()
-        .SetFieldNameMode(NProtobufJson::TJson2ProtoConfig::FieldNameSnakeCaseDense)
-        .SetEnumValueMode(NProtobufJson::TJson2ProtoConfig::EnumCaseInsensetive)
-        .SetAllowUnknownFields(true);
-}
-
-void MergeYamlNodeToProto(const YAML::Node& node, google::protobuf::Message& proto) {
-    if (!node || !node.IsDefined() || node.IsNull()) {
-        return;
-    }
-
-    const NJson::TJsonValue json = NKikimr::NYaml::Yaml2Json(node, true);
-    NProtobufJson::MergeJson2Proto(json, proto, MakeJson2ProtoConfig());
-}
 
 } // namespace
